channels for shop

+++ pizza_order/views.py
@@ -1,5 +1,6 @@
 from django.contrib import messages
 from django.contrib.auth import authenticate, login
+from django.contrib.auth.decorators import login_required
 from django.http import HttpResponse
 from django.views.decorators.csrf import csrf_exempt
 from rest_framework import viewsets
@@ -17,26 +18,45 @@ from .stripe_utils import stripe_session_create, stripe_customer_create
 
 
 class PizzaAdminView(viewsets.ModelViewSet):
-    permission_classes = [IsAdminUser]
+    permission_classes = [IsAuthenticated]
     serializer_class = PizzaSerializer
     queryset = Pizza.objects.all()
 
+    def create(self, request, *args, **kwargs):
+        shop_owner = request.user
+        data = request.data
+        if shop_owner.is_shop_owner:
+            id_shop = data['shop']
+            shop_data = Shop.objects.filter(id=id_shop).first()
+            if shop_data:
+                if shop_owner.id == shop_data.owner.id:
+                    serializer = PizzaSerializer(data=data)
+                    if serializer.is_valid(raise_exception=True):
+                        serializer.save()
+                        return Response({create})
+                    return Response({data_error})
+                return Response({nu})
+            return Response({nd})
+        return Response({nu})
+
     def destroy(self, request, *args, **kwargs):
         data = kwargs['pk']
-        data_pizza = Pizza.objects.get(id=data)
-        pizza_data = CartPizza.objects.filter(pizza_id=data).first()
-        order_pizza = OrderPizza.objects.filter(pizza_id=data).first()
-        if pizza_data or order_pizza:
-            """
-            Here it is checking for pizza id is present or not in cart and order.
-            If present then simply change it to is_deleted = true.
-            If not present then delete it .
-            """
-            data_pizza.is_deleted = True
-            data_pizza.save()
-            return Response({delete})
-        data_pizza.delete()
-        return Response({delete})
+        shop_obj = Shop.objects.prefetch_related('pizza_set').filter(id=data).first()
+        if shop_obj:
+            data_pizza = Pizza.objects.get(id=data)
+            pizza_data = CartPizza.objects.filter(pizza_id=data).first()
+            order_pizza = OrderPizza.objects.filter(pizza_id=data).first()
+            if pizza_data or order_pizza:
+                """
+                Here it is checking for pizza id is present or not in cart and order.
+                If present then simply change it to is_deleted = true.
+                If not present then delete it .
+                """
+                data_pizza.is_deleted = True
+                data_pizza.save()
+                return Response({delete})
+            data_pizza.delete()
+        return Response({nd})
 
     def partial_update(self, request, *args, **kwargs):
         super(PizzaAdminView, self).partial_update(request, *args, **kwargs)
@@ -102,10 +122,38 @@ class CartView(viewsets.ModelViewSet):
         cart, _ = Cart.objects.get_or_create(user=user)
         pizza_id_data = data.get('pizza')
         pizza_data = Pizza.objects.filter(id=pizza_id_data, is_deleted=False).first()
-        if pizza_data:
-            CartPizza.objects.create(pizza_id=data.get('pizza'), cart=cart, quantity=data.get('quantity'))
+        try:
+            cartpizza_data = CartPizza.objects.filter(cart__user=request.user).first()
+        except CartPizza.DoesNotExist:
+            cartpizza_data = None
+        if cartpizza_data:
+            if pizza_data.shop.id == cartpizza_data.shop.id:
+                CartPizza.objects.create(pizza_id=data.get('pizza'), cart=cart, quantity=data.get('quantity'),
+                                         shop_id=pizza_data.shop.id)
+                return Response({create})
+            return Response({nd})
+        else:
+            CartPizza.objects.create(pizza_id=data.get('pizza'), cart=cart, quantity=data.get('quantity'),
+                                     shop_id=pizza_data.shop.id)
             return Response({create})
-        return Response({no_pizza})
+        # if cartpizza_data:
+        #     latest_cart_pizza = cartpizza_data.latest()
+        # cartpizza_data = get_object_or_404(CartPizza, id=pizza_id_data)
+        # pizza_data = Pizza.objects.filter(id=pizza_id_data, is_deleted=False).first()
+        # print(pizza_data.id,cartpizza_data.shop.id)
+        # if cartpizza_data:
+        #     if pizza_data.id == cartpizza_data.shop.id:
+        #         print("3")
+        #         CartPizza.objects.create(pizza_id=data.get('pizza'), cart=cart, quantity=data.get('quantity'),
+        #                                  shop_id=data.get('shop'))
+        #         return Response({create})
+        #     return Response({nd})
+        # elif pizza_data:
+        #     print("2")
+        # CartPizza.objects.create(pizza_id=data.get('pizza'), cart=cart, quantity=data.get('quantity'),
+        #                          shop_id=data.get('shop'))
+        # return Response({create})
+        # return Response({no_pizza})
 
 
 class CartPizzaView(viewsets.ModelViewSet):
@@ -129,13 +177,22 @@ class OrderCreate(viewsets.ModelViewSet):
         address = Address.objects.filter(id=data.get('address'), user=user).first()
         if address:
             cart = Cart.objects.filter(user=user).first()
+            shop_data = cart.cartpizza_set.all().get(cart=cart)
             if cart.total_amount > 0:
                 self.check_object_permissions(request, cart)
                 data['total_amount'] = cart.total_amount
-                serializer = OrderSerializer(data=data, context={"request": request})
+                print("hii")
+                serializer = OrderSerializer(data={**data,"shop": shop_data.shop.id}, context={"request": request})
                 if serializer.is_valid(raise_exception=True):
                     order = serializer.save()
                     mail_send.delay("Order done", order.user.email, "Order of Pizza")
+                    log = {
+                        "order": order.id,
+                        "status": order.status.id
+                    }
+                    status_log = StatusSerializerSignal(data=log)
+                    if status_log.is_valid(raise_exception=True):
+                        status_log.save()
                     for pizza_data in CartPizza.objects.filter(cart=cart):
                         OrderPizza.objects.create(order_id=order.id, pizza=pizza_data.pizza,
                                                   price=pizza_data.pizza.price,
@@ -176,10 +233,12 @@ def login_request(request):
         user = authenticate(email=email, password=password)
         if user:
             login(request, user)
-            if not user.is_delivery_boy:
-                return redirect('home')
+            if user.is_delivery_boy:
+                return redirect('order_delivered')
+            elif user.is_shop_owner:
+                return redirect('shop_owner')
             else:
-                return redirect('order-delivered')
+                return redirect('home')
         else:
             return redirect('/')
     else:
@@ -250,9 +309,12 @@ def order_delivered(request):
     return render(request, 'order_delivered.html')
 
 
-def order_delivered_url(request, id):
+def order_delivered_url(request, id, data):
+    # print(data)
+    # sr = User.objects.filter(id=data).first()
     # DeliveryBoy.objects.create(delivery_boy=request.user,order_id=id)
     order_data = Order.objects.filter(id=id).first()
+    # print(order_data.user.is_delivery_boy)
     delivery_boy = DeliveryBoy.objects.filter(order_id=id).first()
     if not delivery_boy:
         serializer = DeliveryBoySerializer(data={
@@ -264,3 +326,48 @@ def order_delivered_url(request, id):
             order_data.save()
             return HttpResponse(f"Order Accepted of {request.user}")
     return HttpResponse("Order already accepted")
+
+
+class ShopCreate(viewsets.ModelViewSet):
+    permission_classes = [IsOwner]
+    serializer_class = ShopSerializer
+    queryset = Shop.objects.all()
+    lookup_field = 'pk'
+
+    def create(self, request, *args, **kwargs):
+        data = request.data
+        user = request.user
+        if user.is_shop_owner:
+            shop_data = ShopSerializer(data=data, context={
+                'request': request
+            })
+            if shop_data.is_valid(raise_exception=True):
+                shop_data.save()
+                return Response({create})
+            return Response({data_error})
+        return Response({nu})
+
+    def partial_update(self, request, *args, **kwargs):
+        data_id = kwargs['pk']
+        shop_user = Shop.objects.filter(id=data_id).first()
+        self.check_object_permissions(request, shop_user)
+        super(ShopCreate, self).partial_update(self, request, *args, **kwargs)
+        return Response({update})
+
+    def destroy(self, request, *args, **kwargs):
+        data_id = kwargs['pk']
+        shop_user = Shop.objects.filter(id=data_id).first()
+        self.check_object_permissions(request, shop_user)
+        super(ShopCreate, self).destroy(self, request, *args, **kwargs)
+        return Response({delete})
+
+
+@login_required
+def shop_owner(request):
+    # user = User.objects.get(id=request.user.id)
+    # shop_data = user.shop_set.filter(owner=user).first()
+    # context = {
+    #     "shop_data":shop_data
+    # }
+    # print(usr.shop_set.all().filter(owner=usr).first())
+    return render(request, 's_owner.html')
